Moved Logout section code from interventionsVC to MenuVC. Changed logout command closure
