+++ Controllers/InterventionViewController.swift
@@ -574,54 +574,4 @@ class InterventionViewController: UIViewController, UITableViewDelegate, UITable
   @objc private func presentSideMenu(sender: UIButton) {
     performSegue(withIdentifier: "presentSideMenu", sender: sender)
   }
-
-  // MARK: - Logout
-
-  @objc private func presentLogoutAlert() {
-    let alert = UIAlertController(title: "disconnect_prompt".localized, message: nil, preferredStyle: .actionSheet)
-    let cancelAction = UIAlertAction(title: "cancel".localized, style: .cancel, handler: nil)
-    let logoutAction = UIAlertAction(title: "menu_logout".localized, style: .destructive, handler: { action in
-      self.logoutUser()
-    })
-
-    alert.addAction(cancelAction)
-    alert.addAction(logoutAction)
-    present(alert, animated: true)
-  }
-
-  private func logoutUser() {
-    let authentificationService = AuthentificationService(username: "", password: "")
-
-    authentificationService.logout()
-    UserDefaults.standard.removePersistentDomain(forName: Bundle.main.bundleIdentifier!)
-    UserDefaults.standard.synchronize()
-    emptyAllCoreData()
-    navigationController?.popViewController(animated: true)
-  }
-
-  private func emptyAllCoreData() {
-    let entityNames = appDelegate.persistentContainer.managedObjectModel.entities.map({ (entity) -> String in
-      return entity.name!
-    })
-
-    for entityName in entityNames {
-      batchDeleteEntity(name: entityName)
-    }
-  }
-
-  private func batchDeleteEntity(name: String) {
-    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else {
-      return
-    }
-
-    let managedContext = appDelegate.persistentContainer.viewContext
-    let fetch = NSFetchRequest<NSFetchRequestResult>(entityName: name)
-    let request = NSBatchDeleteRequest(fetchRequest: fetch)
-
-    do {
-      try managedContext.execute(request)
-    } catch let error as NSError {
-      print("Could not remove data. \(error), \(error.userInfo)")
-    }
-  }
 }
