+++ Controllers/MenuViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import CoreData
 
 class MenuViewController: UITableViewController {
 
@@ -22,7 +23,7 @@ class MenuViewController: UITableViewController {
 
   private func initializeCommands() {
     let termsCommand = Command(name: "privacy_policy", assetName: "doc", closure: test)
-    let logoutCommand = Command(name: "logout", assetName: "logout", closure: logout)
+    let logoutCommand = Command(name: "logout", assetName: "logout", closure: presentLogoutAlert)
 
     commands = [termsCommand, logoutCommand]
   }
@@ -66,7 +67,54 @@ class MenuViewController: UITableViewController {
     print("cgu")
   }
 
-  private func logout() {
-    print("logout")
+  // MARK: - Logout
+
+  private func presentLogoutAlert() {
+    let alert = UIAlertController(title: "disconnect_prompt".localized, message: nil, preferredStyle: .actionSheet)
+    let cancelAction = UIAlertAction(title: "cancel".localized, style: .cancel, handler: nil)
+    let logoutAction = UIAlertAction(title: "logout".localized, style: .destructive, handler: { action in
+      self.logoutUser()
+    })
+
+    alert.addAction(cancelAction)
+    alert.addAction(logoutAction)
+    present(alert, animated: true)
+  }
+
+  private func logoutUser() {
+    let authentificationService = AuthentificationService(username: "", password: "")
+
+    authentificationService.logout()
+    UserDefaults.standard.removePersistentDomain(forName: Bundle.main.bundleIdentifier!)
+    UserDefaults.standard.synchronize()
+    emptyAllCoreData()
+    navigationController?.popViewController(animated: true)
+  }
+
+  private func emptyAllCoreData() {
+    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else { return }
+    let entityNames = appDelegate.persistentContainer.managedObjectModel.entities.map({ (entity) -> String in
+      return entity.name!
+    })
+
+    for entityName in entityNames {
+      batchDeleteEntity(name: entityName)
+    }
+  }
+
+  private func batchDeleteEntity(name: String) {
+    guard let appDelegate = UIApplication.shared.delegate as? AppDelegate else {
+      return
+    }
+
+    let managedContext = appDelegate.persistentContainer.viewContext
+    let fetch = NSFetchRequest<NSFetchRequestResult>(entityName: name)
+    let request = NSBatchDeleteRequest(fetchRequest: fetch)
+
+    do {
+      try managedContext.execute(request)
+    } catch let error as NSError {
+      print("Could not remove data. \(error), \(error.userInfo)")
+    }
   }
 }
